refactor: 견적서 작성 부분 컴포넌트화
